requirements.txt
requests
cv2
numpy
selenium
moviepy
youtube_dl
yt_dlp
pynput
phonenumbers
geocoder
pyzipper
beautifulsoup4
pyautogui
pybluez
telethon
tensorflow
joblib
scapy
paramiko
pillow
exifread
scikit-learn
paramiko